feat: OData 'in' operator, bbox spatial shorthand, and $format=parquet parity (#1621)
